ollama_faiss_flask_app/requirements.txt
faiss-cpu
sentence-transformers
requests
ollama
pdfplumber
scholarly
flask